Make our own copy of EncoderStack.
EncoderStack was from: https://github.com/tensorflow/models/blob/master/official/legacy/transformer/transformer.py, and attention_layer.py and ffn_layer.py are from the same directory.

The reason why we're making our own copies is because these components are now part of "legacy" in the tensorflow_models tool, so it's better for us to make a copy and maintain them on our own.

PiperOrigin-RevId: 435112609